// src/app/modules/cleansing/models/officer-details.model.ts
export interface ICeoDetails {
  ceoAddress: string;
  ceoAppDate: string;
  ceoCountry: string;
  ceoFatherName: string;
  ceoNIC: string;
  ceoName: string;
  compIncNo: string;
  key: string;
  message: string;
  timestamp: string;
  userId: string;
}

export interface IChiefDetails {
  cacAddress: string;
  cacAppDate: string;
  cacCountry: string;
  cacFatherName: string;
  cacNIC: string;
  cacName: string;
  compIncNo: string;
  key: string;
  message: string;
  timestamp: string;
  userId: string;
}

export interface IAdvisorDetails {
  compIncNo: string;
  key: string;
  lglAdvAddress: string;
  lglAdvAppDate: string;
  lglAdvCountry: string;
  lglAdvFatherName: string;
  lglAdvNIC: string;
  lglAdvName: string;
  message: string;
  timestamp: string;
}

export interface IAgentDetails {
  compIncNo: string;
  key: string;
  message: string;
  mngAddress: string;
  mngAppDate: string;
  mngCountry: string;
  mngFatherName: string;
  mngNIC: string;
  mngName: string;
  timestamp: string;
}

export interface ISecretaryDetails {
  compIncNo: string;
  key: string;
  message: string;
  secrAddress: string;
  secrAppDate: string;
  secrCountry: string;
  secrFatherName: string;
  secrNIC: string;
  secrName: string;
  timestamp: string;
}